Update claudeContentScript.js
Claude completely changed how it's web interface works and the previous version didn't work at all, so I had to completely rewrite this.. It is not perfect.  I streams in inline code blocks, but does not stream in artifacts yet.. For code artifacts, you must click on the artifact after it is finished to bring the code into the editor.